% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Clamp Numeric Values}
\usage{
clamp(x, lower, upper)
}
\arguments{
\item{x}{Numeric vector.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}
}
\description{
Clamp Numeric Values
}
