% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_and_msm.R
\name{fit_msm_gee_cont}
\alias{fit_msm_gee_cont}
\title{Fit GEE MSM for Continuous Outcomes}
\usage{
fit_msm_gee_cont(df_w)
}
\arguments{
\item{df_w}{Weighted data frame.}
}
\description{
Fit GEE MSM for Continuous Outcomes
}
