% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_suggested}
\alias{require_suggested}
\title{Check That a Suggested Package Is Installed}
\usage{
require_suggested(pkg)
}
\arguments{
\item{pkg}{Character string: package name.}
}
\value{
Invisible NULL. Called for its side effect of raising an error.
}
\description{
Check That a Suggested Package Is Installed
}
\keyword{internal}
