% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanitize_nl_col_name}
\alias{sanitize_nl_col_name}
\title{Sanitize Non-linear Term Column Name}
\usage{
sanitize_nl_col_name(term_str)
}
\arguments{
\item{term_str}{Non-linear expression or interaction string.}
}
\value{
Sanitized column name starting with \code{"NL_"}.
}
\description{
Sanitize Non-linear Term Column Name
}
\keyword{internal}
