% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Required Columns Exist in a Data Frame}
\usage{
validate_columns(df, cols, fn_name = sys.call(sys.parent())[[1]])
}
\arguments{
\item{df}{Data frame to check.}

\item{cols}{Character vector of required column names.}

\item{fn_name}{Name of the calling function (for the error message).}
}
\value{
Invisible NULL. Called for its side effect of raising an error.
}
\description{
Validate Required Columns Exist in a Data Frame
}
\keyword{internal}
