% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_min_rows}
\alias{validate_min_rows}
\title{Validate Data Frame Has Sufficient Rows}
\usage{
validate_min_rows(df, min_rows, fn_name = sys.call(sys.parent())[[1]])
}
\arguments{
\item{df}{Data frame.}

\item{min_rows}{Minimum number of rows required.}

\item{fn_name}{Name of the calling function.}
}
\value{
Invisible NULL.
}
\description{
Validate Data Frame Has Sufficient Rows
}
\keyword{internal}
