% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_and_msm.R
\name{compute_stabilized_iptw}
\alias{compute_stabilized_iptw}
\title{Compute Stabilized IPTW for Binary Compliance}
\usage{
compute_stabilized_iptw(
  df,
  c_var = "Compliant",
  ps_var = "ps_hat",
  visit_var = "VisitNumber",
  truncate = c(0.01, 0.99)
)
}
\arguments{
\item{df}{Data frame with columns: Compliant, VisitNumber, ps_hat.}

\item{c_var}{Name of compliance variable.}

\item{ps_var}{Name of propensity score variable.}

\item{visit_var}{Name of visit number variable.}

\item{truncate}{Trimming quantiles.}
}
\description{
Compute Stabilized IPTW for Binary Compliance
}
