% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_and_msm.R
\name{fit_msm_cox}
\alias{fit_msm_cox}
\title{Fit Cox MSM (Start-Stop Intervals) with Robust SEs}
\usage{
fit_msm_cox(tte_ivl_w)
}
\arguments{
\item{tte_ivl_w}{Weighted TTE interval data.}
}
\description{
Fit Cox MSM (Start-Stop Intervals) with Robust SEs
}
