% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_and_msm.R
\name{fit_msm_gee_bin}
\alias{fit_msm_gee_bin}
\title{Fit GEE MSM for Binary Outcomes}
\usage{
fit_msm_gee_bin(df_w)
}
\arguments{
\item{df_w}{Weighted data frame.}
}
\description{
Fit GEE MSM for Binary Outcomes
}
