% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_patient_data_demographic}
\alias{generate_patient_data_demographic}
\title{Generate Demographics}
\usage{
generate_patient_data_demographic(
  num_patients = 1000,
  seed = 123,
  n_cohorts = 3,
  age_range = c(18, 90),
  gender_levels = c("Male", "Female"),
  treatment_prefix = "Treatment_"
)
}
\arguments{
\item{num_patients}{Total number of patients.}

\item{seed}{Random seed.}

\item{n_cohorts}{Number of treatment cohorts.}

\item{age_range}{Range of patient ages.}

\item{gender_levels}{Levels for gender variable.}

\item{treatment_prefix}{Prefix for treatment names.}
}
\description{
Generate Demographics
}
