% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of probabilities.
}
\description{
Computes the inverse logit (logistic) transformation.
}
\examples{
inv_logit(0)
inv_logit(c(-2, 0, 2))
}
