% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_tidymodels.R
\name{ps_model_spec}
\alias{ps_model_spec}
\title{Model spec factory for PS}
\usage{
ps_model_spec(model = c("glm", "xgboost", "ranger"), tune = FALSE)
}
\arguments{
\item{model}{One of "glm","xgboost","ranger".}

\item{tune}{Logical; if TRUE, uses tune() placeholders for parameters.}
}
\description{
Model spec factory for PS
}
