% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_tidymodels.R
\name{ps_recipe}
\alias{ps_recipe}
\title{Build a PS Recipe}
\usage{
ps_recipe(
  df,
  outcome = "Compliant",
  formula =
    as.formula("Compliant ~ Treatment + VisitNumber + Age + GenderMale + SideEffect + Biomarker")
)
}
\arguments{
\item{df}{Data frame.}

\item{outcome}{Name of binary outcome column ("Compliant").}

\item{formula}{Predictors formula (default includes Treatment, time, demos, SE, Biomarker).}
}
\description{
Build a PS Recipe
}
